+++ src/Manopt.jl
@@ -288,6 +288,7 @@ export get_proximal_map,
     get_dual_prox!,
     get_differential_dual_prox,
     get_differential_dual_prox!,
+    get_stopping_criterion,
     set_gradient!,
     set_iterate!,
     set_manopt_parameter!,
@@ -404,6 +405,7 @@ export ExactPenaltyCost, ExactPenaltyGrad, AugmentedLagrangianCost, AugmentedLag
 export Stepsize
 export ArmijoLinesearch,
     ConstantStepsize, DecreasingStepsize, Linesearch, NonmonotoneLinesearch
+export AdaptiveWNGradient
 export get_stepsize, get_initial_stepsize, get_last_stepsize
 #
 # Stopping Criteria
@@ -428,7 +430,8 @@ export StopAfter,
     StopWhenSmallerOrEqual,
     StopWhenStepsizeLess,
     StopWhenTrustRegionIsExceeded
-export get_active_stopping_criteria, get_stopping_criteria, get_reason
+export get_active_stopping_criteria,
+    get_stopping_criteria, get_reason, get_stopping_criterion
 export update_stopping_criterion!
 #
 # Data functions