+++ src/plans/objective.jl
@@ -105,12 +105,12 @@ If `recursive` is set to `false`, only the most outer decorator is taken away in
 function get_objective(o::AbstractManifoldObjective, recursive=true)
     return _get_objective(o, dispatch_objective_decorator(o), recursive)
 end
-_get_objective(o::AbstractManifoldObjective, ::Val{false}, rec) = o
-function _get_objective(o::AbstractManifoldObjective, ::Val{true}, rec)
+_get_objective(o::AbstractManifoldObjective, ::Val{false}, rec=true) = o
+function _get_objective(o::AbstractManifoldObjective, ::Val{true}, rec=true)
     return rec ? get_objective(o.objective) : o.objective
 end
 function status_summary(o::AbstractManifoldObjective{E}) where {E}
-    return "$(nameof(typeof(o))){$E}"
+    return ""#"$(nameof(typeof(o))){$E}"
 end
 # Default undecorate for summary
 function status_summary(co::AbstractDecoratedManifoldObjective)
@@ -126,11 +126,9 @@ function show(io::IO, co::AbstractDecoratedManifoldObjective)
 end
 
 function show(io::IO, t::Tuple{<:AbstractManifoldObjective,P}) where {P}
+    s = "$(status_summary(t[1]))"
+    length(s) > 0 && (s = "$(s)\n\n")
     return print(
-        io,
-        """
-$(status_summary(t[1]))
-
-To access the solver result, call `get_solver_result` on this variable.""",
+        io, "$(s)To access the solver result, call `get_solver_result` on this variable."
     )
 end