+++ src/plans/stepsize.jl
@@ -45,13 +45,14 @@ end
 A functor that always returns a fixed step size.
 
 # Fields
-* `length` – constant value for the step size.
+* `length` – constant value for the step size
+* `type` - a symbol that indicates whether the stepsize is relatively (to the gradient norm) or absolutely constant
 
 # Constructors
 
-    ConstantStepsize(s::Real)
+    ConstantStepsize(s::Real, t::Symbol)
 
-initialize the stepsize to a constant `s`.
+initialize the stepsize to a constant `s` of type `t`.
 
     ConstantStepsize(M::AbstractManifold=DefaultManifold(2); stepsize=injectivity_radius(M)/2)
 
@@ -60,20 +61,32 @@ radius, unless the radius is infinity, then the default step size is `1`.
 """
 mutable struct ConstantStepsize{T} <: Stepsize
     length::T
+    type::Symbol
 end
 function ConstantStepsize(
     M::AbstractManifold=DefaultManifold(2);
     stepsize=isinf(injectivity_radius(M)) ? 1.0 : injectivity_radius(M) / 2,
+    type=:relative,
 )
-    return ConstantStepsize{typeof(stepsize)}(stepsize)
+    return ConstantStepsize{typeof(stepsize)}(stepsize, type)
+end
+function ConstantStepsize(stepsize::T) where {T<:Number}
+    return ConstantStepsize{T}(stepsize, :relative)
 end
 function (cs::ConstantStepsize)(
-    ::AbstractManoptProblem, ::AbstractManoptSolverState, ::Any, args...; kwargs...
+    amp::AbstractManoptProblem, ams::AbstractManoptSolverState, ::Any, args...; kwargs...
 )
-    return cs.length
+    s = cs.length
+    if cs.type == :absolute
+        ns = norm(get_manifold(amp), get_iterate(sgs), get_subgradient(sgs))
+        if ns > eps(eltype(s))
+            s /= ns
+        end
+    end
+    return s
 end
 get_initial_stepsize(s::ConstantStepsize) = s.length
-show(io::IO, cs::ConstantStepsize) = print(io, "ConstantStepsize($(cs.length))")
+show(io::IO, cs::ConstantStepsize) = print(io, "ConstantStepsize($(cs.length), $(cs.type))")
 
 @doc raw"""
     DecreasingStepsize()
@@ -87,6 +100,7 @@ A functor that represents several decreasing step sizes
 * `exponent` – (`1`) a value ``e`` the current iteration numbers ``e``th exponential
   is taken of
 * `shift` – (`0`) shift the denominator iterator ``i`` by ``s```.
+* `type` - a symbol that indicates whether the stepsize is relatively (to the gradient norm) or absolutely constant
 
 In total the complete formulae reads for the ``i``th iterate as
 
@@ -98,7 +112,7 @@ and hence the default simplifies to just ``s_i = \frac{l}{i}``
 
 # Constructor
 
-    DecreasingStepsize(l=1,f=1,a=0,e=1,s=0)
+    DecreasingStepsize(l=1,f=1,a=0,e=1,s=0,type=:relative)
 
 Alternatively one can also use the following keyword.
 
@@ -115,8 +129,11 @@ mutable struct DecreasingStepsize <: Stepsize
     subtrahend::Float64
     exponent::Float64
     shift::Int
-    function DecreasingStepsize(l::Real, f::Real=1.0, a::Real=0.0, e::Real=1.0, s::Int=0)
-        return new(l, f, a, e, s)
+    type::Symbol
+    function DecreasingStepsize(
+        l::Real, f::Real=1.0, a::Real=0.0, e::Real=1.0, s::Int=0, type::Symbol=:relative
+    )
+        return new(l, f, a, e, s, type)
     end
 end
 function DecreasingStepsize(
@@ -126,13 +143,21 @@ function DecreasingStepsize(
     subtrahend=0.0,
     exponent=1.0,
     shift=0,
+    type::Symbol=:relative,
 )
-    return DecreasingStepsize(length, factor, subtrahend, exponent, shift)
+    return DecreasingStepsize(length, factor, subtrahend, exponent, shift, type)
 end
 function (s::DecreasingStepsize)(
-    ::P, ::O, i::Int, args...; kwargs...
+    amp::P, ams::O, i::Int, args...; kwargs...
 ) where {P<:AbstractManoptProblem,O<:AbstractManoptSolverState}
-    return (s.length - i * s.subtrahend) * (s.factor^i) / ((i + s.shift)^(s.exponent))
+    ds = (s.length - i * s.subtrahend) * (s.factor^i) / ((i + s.shift)^(s.exponent))
+    if s.type == :absolute
+        ns = norm(get_manifold(amp), get_iterate(sgs), get_subgradient(sgs))
+        if ns > eps(eltype(ds))
+            ds /= ns
+        end
+    end
+    return ds
 end
 get_initial_stepsize(s::DecreasingStepsize) = s.length
 function show(io::IO, s::DecreasingStepsize)
@@ -957,6 +982,172 @@ function status_summary(a::WolfePowellBinaryLinesearch)
     s = (a.last_stepsize > 0) ? "\nand the last stepsize used was $(a.last_stepsize)." : ""
     return "$a$s"
 end
+
+@doc raw"""
+    AdaptiveWNGradient <: DirectionUpdateRule
+
+Represent an adaptive gradient method introduced by [^GrapigliaStella2023].
+
+Given a positive threshold ``\hat c \mathbb N``,
+an minimal bound ``b_{\mathrm{min}} > 0``,
+an initial ``b_0 ≥ b_{\mathrm{min}}``, and a
+gradient reduction factor threshold ``\alpha \in [0,1).
+
+Set ``c_0=0`` and use ``\omega_0 = \lVert \operatorname{grad} f(p_0) \rvert_{p_0}``.
+
+For the first iterate we use the initial step size ``s_0 = \frac{1}{b_0}``
+
+Then, given the last gradient ``X_{k-1} = \operatorname{grad} f(x_{k-1})``,
+and a previous ``\omega_{k-1}``, the values ``(b_k, \omega_k, c_k)`` are computed
+using ``X_k = \operatorname{grad} f(p_k)`` and the following cases
+
+If ``\lVert X_k \rVert_{p_k} \leq \alpha\omega_{k-1}``, then let
+``\hat b_{k-1} \in [b_\mathrm{min},b_{k-1}]`` and set
+
+```math
+(b_k, \omega_k, c_k) = \begin{cases}
+\bigl(\hat b_{k-1}, \lVert X_k\rVert_{p_k}, 0 \bigr) & \text{ if } c_{k-1}+1 = \hat c\\
+\Bigl(b_{k-1} + \frac{\lVert X_k\rVert_{p_k}^2}{b_{k-1}}, \omega_{k-1}, c_{k-1}+1 \Bigr) & \text{ if } c_{k-1}+1<\hat c
+\end{cases}
+```
+
+If ``\lVert X_k \rVert_{p_k} > \alpha\omega_{k-1}``, the set
+
+```math
+(b_k, \omega_k, c_k) =
+\Bigl( b_{k-1} + \frac{\lVert X_k\rVert_{p_k}^2}{b_{k-1}}, \omega_{k-1}, 0)
+```
+
+and return the step size ``s_k = \frac{1}{b_k}``.
+
+Note that for ``α=0`` this is the Riemannian variant of WNGRad
+
+# Fields
+* `count_threshold::Int` (`4`) an `Integer` for ``\hat c``
+* `minimal_bound::Float64` (`1e-4`) for ``b_{\mathrm{min}}``
+* `alternate_bound::Function` (`(bk, hat_c) -> min(gradient_bound, max(gradient_bound, bk/(3*hat_c)`)
+  how to determine ``\hat b_k`` as a function of `(bmin, bk, hat_c) -> hat_bk`
+* `gradient_reduction::Float64` (`0.9`)
+* `gradient_bound` `norm(M, p0, grad_f(M,p0))` the bound ``b_k``.
+
+as well as the internal fields
+
+* `weight` for ``ω_k`` initialised to ``ω_0 = ```norm(M, p0, grad_f(M,p0))` if this is not zero, `1.0` otherwise.
+* `count` for the ``c_k``, initialised to ``c_0 = 0``.
+
+# Constructor
+
+    AdaptiveWNGrad(M=DefaultManifold, grad_f=(M,p) -> zero_vector(M,rand(M)), p=rand(M); kwargs...)
+
+Where all above fields with defaults are keyword arguments.
+An additional keyword arguments
+
+* `adaptive` (`true`) switches the `gradient_reduction ``α`` to `0`.
+* `evaluation` (`AllocatingEvaluation()`) specifies whether the gradient (that is used for initialisation only) is mutating or allocating
+
+[^GrapigliaStella2023]:
+    > Grapiglia, G. N., Stella, G. F. D.:
+    > An Adaptive Riemannian Gradient Method Without Function Evaluations
+    > Journal of Optimization Theory and Applications (197), pp. 1140–1160, 2023.
+    > doi: [10.1007/s10957-023-02227-y](https://doi.org/10.1007/s10957-023-02227-y),
+    > preprint: [optimization-online.org/wp-content/uploads/2022/04/8864.pdf](https://optimization-online.org/wp-content/uploads/2022/04/8864.pdf)
+"""
+mutable struct AdaptiveWNGradient{I<:Integer,R<:Real,F<:Function} <: Stepsize
+    count_threshold::I
+    minimal_bound::R
+    alternate_bound::F
+    gradient_reduction::R
+    gradient_bound::R
+    weight::R
+    count::I
+end
+function AdaptiveWNGradient(
+    M::AbstractManifold=DefaultManifold(),
+    (grad_f!!)=(M, p) -> zero_vector(M, rand(M)),
+    p=rand(M);
+    evaluation::E=AllocatingEvaluation(),
+    adaptive::Bool=true,
+    count_threshold::I=4,
+    minimal_bound::R=1e-4,
+    gradient_reduction::R=adaptive ? 0.9 : 0.0,
+    gradient_bound::R=norm(
+        M,
+        p,
+        if evaluation == AllocatingEvaluation()
+            grad_f!!(M, p)
+        else
+            grad_f!!(M, zero_vector(M, p), p)
+        end,
+    ),
+    alternate_bound::F=(bk, hat_c) -> min(
+        gradient_bound == 0 ? 1.0 : gradient_bound, max(minimal_bound, bk / (3 * hat_c))
+    ),
+    kwargs...,
+) where {I<:Integer,R<:Real,F<:Function,E<:AbstractEvaluationType}
+    if gradient_bound == 0
+        # If the gradient bound defaults to zero, we set it to 1
+        gradient_bound = 1.0
+    end
+    return AdaptiveWNGradient{I,R,F}(
+        count_threshold,
+        minimal_bound,
+        alternate_bound,
+        gradient_reduction,
+        gradient_bound,
+        gradient_bound,
+        0,
+    )
+end
+function (awng::AdaptiveWNGradient)(
+    mp::AbstractManoptProblem, s::AbstractGradientSolverState, i, args...; kwargs...
+)
+    M = get_manifold(mp)
+    p = get_iterate(s)
+    X = get_gradient(mp, p)
+    isnan(awng.weight) || (awng.weight = norm(M, p, X)) # init ω_0
+    if i == 0 # init fields
+        awng.weight = norm(M, p, X) # init ω_0
+        (awng.weight == 0) && (awng.weight = 1.0)
+        awng.count = 0
+        return 1 / awng.gradient_bound
+    end
+    grad_norm = norm(M, p, X)
+    if grad_norm < awng.gradient_reduction * awng.weight # grad norm < αω_{k-1}
+        if awng.count + 1 == awng.count_threshold
+            awng.gradient_bound = awng.alternate_bound(
+                awng.gradient_bound, awng.count_threshold
+            )
+            awng.weight = grad_norm
+            awng.count = 0
+        else
+            awng.gradient_bound = awng.gradient_bound + grad_norm^2 / awng.gradient_bound
+            #weight stays unchanged
+            awng.count += 1
+        end
+    else
+        awng.gradient_bound = awng.gradient_bound + grad_norm^2 / awng.gradient_bound
+        #weight stays unchanged
+        awng.count = 0
+    end
+    return 1 / awng.gradient_bound
+end
+get_initial_stepsize(awng::AdaptiveWNGradient) = 1 / awng.gradient_bound
+get_last_stepsize(awng::AdaptiveWNGradient) = 1 / awng.gradient_bound
+function show(io::IO, awng::AdaptiveWNGradient)
+    s = """
+    AdaptiveWNGradient(;
+      count_threshold=$(awng.count_threshold),
+      minimal_bound=$(awng.minimal_bound),
+      alternate_bound=$(awng.alternate_bound),
+      gradient_reduction=$(awng.gradient_reduction),
+      gradient_bound=$(awng.gradient_bound)
+    )
+
+    as well as internally the weight ω_k = $(awng.weight) and current count c_k = $(awng.count).
+    """
+    return print(io, s)
+end
+
 @doc raw"""
     get_stepsize(amp::AbstractManoptProblem, ams::AbstractManoptSolverState, vars...)