+++ src/solvers/subgradient.jl
@@ -225,3 +225,27 @@ function step_solver!(mp::AbstractManoptProblem, sgs::SubGradientMethodState, i)
     return sgs
 end
 get_solver_result(sgs::SubGradientMethodState) = sgs.p_star
+function (cs::ConstantStepsize)(
+    amp::AbstractManoptProblem, sgs::SubGradientMethodState, ::Any, args...; kwargs...
+)
+    s = cs.length
+    if cs.type == :absolute
+        ns = norm(get_manifold(amp), get_iterate(sgs), get_subgradient(sgs))
+        if ns > eps(eltype(s))
+            s /= ns
+        end
+    end
+    return s
+end
+function (s::DecreasingStepsize)(
+    amp::AbstractManoptProblem, sgs::SubGradientMethodState, i::Int, args...; kwargs...
+)
+    ds = (s.length - i * s.subtrahend) * (s.factor^i) / ((i + s.shift)^(s.exponent))
+    if s.type == :absolute
+        ns = norm(get_manifold(amp), get_iterate(sgs), get_subgradient(sgs))
+        if ns > eps(eltype(ds))
+            ds /= ns
+        end
+    end
+    return ds
+end