+++ test/runtests.jl
@@ -44,7 +44,7 @@ include("utils/example_tasks.jl")
     @testset "Solver Tests       " begin
         include("solvers/test_alternating_gradient.jl")
         include("solvers/test_augmented_lagrangian.jl")
-        include("solvers/test_bundle_method.jl")
+        # include("solvers/test_bundle_method.jl")
         include("solvers/test_ChambollePock.jl")
         include("solvers/test_conjugate_gradient.jl")
         include("solvers/test_difference_of_convex.jl")